[feature] 作业提交

+++ ning-user-web/src/main/java/com/ning/geekbang/user/web/repository/UserRepository.java
@@ -8,7 +8,7 @@ public interface UserRepository {
 
     boolean save(User user);
 
-    boolean deleteById(Long userId);
+    boolean deleteById(Integer userId);
 
     boolean update(User user);
